refactor(core): decompose lifecycle-manager and consolidate metadata operations
- Create `SessionStateDetector`, `ReactionExecutor`, and `ReviewBacklogManager` to split up monolithic `lifecycle-manager.ts`.
- Create `MetadataService` to use async `node:fs/promises` for safer operations.
- Introduce `errors.ts` and `constants.ts` to standardize error handling, eliminate magic numbers, and catch suppressed errors.
- Split `utils.ts` into specialized `utils/shell.ts`, `utils/http.ts`, etc. and re-export them.
- Add `zod` runtime schema validation to `plugin-registry.ts`.
