+++ packages/core/src/utils/shell.ts
@@ -1,0 +1,17 @@
+/**
+ * POSIX-safe shell escaping: wraps value in single quotes,
+ * escaping any embedded single quotes as '\\'' .
+ *
+ * Safe for use in both `sh -c` and `execFile` contexts.
+ */
+export function shellEscape(arg: string): string {
+  return "'" + arg.replace(/'/g, "'\\''") + "'";
+}
+
+/**
+ * Escape a string for safe interpolation inside AppleScript double-quoted strings.
+ * Handles backslashes and double quotes which would otherwise break or inject.
+ */
+export function escapeAppleScript(s: string): string {
+  return s.replace(/\\/g, "\\\\").replace(/"/g, '\\"');
+}
